+++ backend/data_tools/tests/etl_data_from_excel/test_utils.py
@@ -1,4 +1,4 @@
-from data_tools.src.etl_data_from_excel.utils import clean_rows
+from data_tools.src.pipeline_data_from_excel.utils import clean_rows
 
 
 def test_clean_rows_happy():
